backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
PyGithub==1.59.1
python-dotenv==1.0.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
supabase==2.3.0
postgrest==0.13.2
